fdfdas

+++ internal/chain/trading.go
@@ -161,7 +161,10 @@ func PoolID(d map[string]any) (string, error) {
 		}
 		args = append(args, abi.Argument{Type: t})
 	}
-	b, err := args.Pack(c0, c1, uint64(ToBig(d["fee"]).Uint64()), ToBig(d["tickSpacing"]).Int64(), hooks)
+	// ABI uint24/int24 values use the native Go widths used by the V4
+	// structs (uint32/int32). Passing uint64/int64 makes go-ethereum reject
+	// the value with "cannot use uint64 as type ptr as argument".
+	b, err := args.Pack(c0, c1, uint32(ToBig(d["fee"]).Uint64()), int32(ToBig(d["tickSpacing"]).Int64()), hooks)
 	if err != nil {
 		return "", err
 	}
